Update read and write
