+++ raftstore/integration_basic_test.go
@@ -14,10 +14,12 @@
 package raftstore
 
 import (
+	"fmt"
 	"testing"
 
 	"github.com/matrixorigin/matrixcube/config"
 	"github.com/matrixorigin/matrixcube/util/leaktest"
+	"github.com/stretchr/testify/assert"
 )
 
 func TestSingleTestClusterStartAndStop(t *testing.T) {
@@ -86,4 +88,50 @@ func TestClusterStartConcurrent(t *testing.T) {
 
 	c.WaitShardByCountPerNode(1, testWaitTimeout)
 	c.CheckShardCount(1)
+}
+
+func TestReadAndWriteAndRestart(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping in short mode.")
+		return
+	}
+
+	defer leaktest.AfterTest(t)()
+
+	fn := func(n int) {
+		c := NewTestClusterStore(t,
+			WithTestClusterNodeCount(n),
+			DiskTestCluster)
+
+		c.Start()
+		defer c.Stop()
+
+		c.WaitShardByCountPerNode(1, testWaitTimeout)
+		c.WaitLeadersByCount(1, testWaitTimeout)
+		c.CheckShardCount(1)
+
+		kv := c.CreateTestKVClient(0)
+		defer kv.Close()
+
+		for i := 0; i < 1; i++ {
+			assert.NoError(t, kv.Set(fmt.Sprintf("k-%d", i), fmt.Sprintf("v-%d", i), testWaitTimeout))
+		}
+
+		c.Restart()
+		c.WaitShardByCountPerNode(1, testWaitTimeout)
+		c.WaitLeadersByCount(1, testWaitTimeout)
+		c.CheckShardCount(1)
+
+		kv2 := c.CreateTestKVClient(0)
+		defer kv2.Close()
+
+		for i := 0; i < 1; i++ {
+			v, err := kv2.Get(fmt.Sprintf("k-%d", i), testWaitTimeout)
+			assert.NoError(t, err)
+			assert.Equal(t, fmt.Sprintf("v-%d", i), v)
+		}
+	}
+
+	// fn(1)
+	fn(3)
 }
